requirements.txt
fastapi>=0.115
uvicorn>=0.34
requests>=2.31
anthropic>=0.49
twilio>=9.0
aiosqlite>=0.20
pydantic-settings>=2.0
python-multipart>=0.0.9
